vocoeng-django-frontend/requirements.txt
# Django framework and REST API
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database connectivity
psycopg2-binary==2.9.7
django-environ==0.11.2

# Google Cloud dependencies
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
google-cloud-firestore==2.11.1
google-cloud-sql-connector==1.4.3

# Production WSGI server
gunicorn==21.2.0

# Admin interface enhancements
django-admin-interface==0.25.0
django-colorfield==0.7.2

# API documentation
drf-spectacular==0.26.5

# Authentication and security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.4

# Data processing and analytics
pandas==2.1.3
numpy==1.24.3

# HTTP client for external API calls
requests==2.31.0

# Logging and monitoring
structlog==23.1.0
django-structlog==4.0.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0

# Type hints and validation
pydantic==2.4.2

# Async processing
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
